+++ src/test/java/nextstep/AcceptanceTest.java
@@ -1,4 +1,4 @@
-package nextstep.subway.acceptance;
+package nextstep;
 
 import io.restassured.RestAssured;
 import nextstep.subway.utils.DatabaseCleanup;
