+++ src/test/java/nextstep/auth/unit/SessionAuthenticationConverterTest.java
@@ -1,8 +1,8 @@
-package nextstep.subway.unit;
+package nextstep.auth.unit;
 
 import nextstep.auth.authentication.AuthenticationToken;
 import nextstep.auth.authentication.session.SessionAuthenticationConverter;
-import nextstep.subway.support.MockRequest;
+import nextstep.support.MockRequest;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 import org.springframework.mock.web.MockHttpServletRequest;
